requirements.txt
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
google-api-core==1.26.3
google-api-python-client==2.4.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.1
oauthlib==3.1.0
packaging==20.9
protobuf==3.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
tqdm==4.60.0
uritemplate==3.0.1
urllib3==1.26.4
